extract Ledger Query Service Conf from test_lib.js
